[CI] Include triton backend and online serving benchmark into CI (#1408)
